스탯, 스킬레벨 정보 저장
TODO : 장비 정보 불러오기

+++ Slime_Clicker_Project/Assets/3.Scripts/UI/UI_StatUpgrade.cs
@@ -26,9 +26,10 @@ public class UI_StatUpgrade : RootUI
         Managers.Instance.StatUpgrade.OnStatChanged += OnStatChanged;
         var initialStats = Managers.Instance.StatUpgrade.GetStatInfo(statType);
 
-        UpdateStat(initialStats.level, (int)initialStats.bonus, initialStats.cost);
+        UpdateStat(initialStats.level, initialStats.bonus, initialStats.cost);
         BindEventToObjects();
     }
+
 
 
     private void OnStatChanged(string type, int level, float bonus, int cost)
@@ -48,6 +49,8 @@ public class UI_StatUpgrade : RootUI
             }
         }
     }
+
+
     private void UpdateStat(int level, float bonus, int cost)
     {
         StatLevel.text = $"LV.{level}";
